rename files to match helper names (then it'll work with Webpack for client-side templating)
