+++ src/helpers/ifAll.js
