+++ src/helpers/ifEquals.js
