+++ src/helpers/removeImageTags.js
